Add AudiobookCandidate dataclass and refactor Importer logic
- Introduced AudiobookCandidate dataclass to encapsulate audiobook metadata.
- Updated detect.py to import and utilize the new AudiobookCandidate class.
- Refactored the Importer class to process audiobook candidates more efficiently and check for existing entries in the database.
- Enhanced logging for better tracking of audiobook processing steps.
- Modified audiobook_edit.html to display audiobook title and input path dynamically.
