+++ src/apps/importer/main.py
@@ -1,6 +1,8 @@
 import os
 from loguru import logger
 from src.apps.importer.detect import find_audiobooks
+from src.apps.importer.dataclasses import AudiobookCandidate
+from src.apps.core.models import Audiobook
 
 logger.add("importer.log", rotation="100 MB", retention="10 days", enqueue=True)
 
@@ -16,13 +18,9 @@ def import_audiobooks():
 class Importer:
     def __init__(self):
         self.input_path = "/app/input"
+        self.output_path = "/app/output"
 
-    def run(self):
-        logger.info("Running Importer")
-        logger.info("Checking for new audiobooks in /app/input")
-        candidates = find_audiobooks(self.input_path)
-        logger.info(f"Found {len(candidates)} candidates")
-        for candidate in candidates:
+    def process_candidate(self, candidate: AudiobookCandidate):
             logger.info(f"Processing {candidate.path}")
             logger.info(f"Author: {candidate.author_guess}")
             logger.info(f"Title: {candidate.title_guess}")
@@ -32,4 +30,25 @@ class Importer:
             logger.info(f"Has Cover: {candidate.has_cover}")
             logger.info(f"Track Count: {candidate.track_count}")
             logger.info(f"Total Audio Bytes: {candidate.total_audio_bytes}")
+            # Check if the audiobook already exists
+            if Audiobook.objects.filter(input_path=candidate.path).exists():
+                logger.info(f"Audiobook already exists: {candidate.path}")
+                return
+            Audiobook.objects.create(
+                title=candidate.title_guess,
+                author=candidate.author_guess,
+                input_path=candidate.path,
+                output_path=self.output_path,
+                status="pending",
+            )
+
+    def run(self):
+        logger.info("Running Importer")
+        logger.info("Checking for new audiobooks in /app/input")
+        candidates = find_audiobooks(self.input_path)
+        logger.info(f"Found {len(candidates)} candidates")
+        for candidate in candidates:
+            self.process_candidate(candidate)
+
+
