+++ templates/core/audiobook_edit.html
@@ -5,7 +5,7 @@
 {% block content %}
 <div class="row mb-4">
     <div class="col">
-        <h1>Edit Audiobook</h1>
+        <h1>Edit Audiobook | {{ audiobook.title }} {{audiobook.input_path}}</h1>
     </div>
 </div>
